got the index going.
So I’m using Jade and SASS, but since Harp does not work when node is
installed --without-npm, Harp simply refuses to attempt to run with the
version NVM has installed. Stupid. So I’m just going to be writing in
Jade and SASS, compiling locally, and using the plain HTML and CSS as
the official site language.