resolved padding in list
- added key for San Francisco Display and San Francisco Text fonts by
Apple
- Helvetica Neue will be cached in CSS for resolves where SFD and SFT
cannot be found.
- Nav bar content no longer stretches the length of the entire nav bar.
It is now in a constraint of 1200px. It is literally in a div called
‘constraints’.
- The middle-most items in the list will be dimmed out. Anything in the
nav bar you don’t want dimmed, simply assign the element the class
‘undim’ and it will take effect immediately.
- Font weight is no longer defaulted to 300 as per new typography
rulesets. I will have to write up a new ruleset. But in the meantime,
no header node will have padding or margins.
- Flexbox is now in its own mixin in order to simplify SASS code. Will
later integrate Grunt to compile for me on save instead of me having to
leave SASS and Jade running in the terminal constantly.