pulled San Francisco font
`You may use this Apple Font only for the purposes described in this
License and only if you are a Registered Apple Developer who is part of
an iOS Developer Program team, or as otherwise expressly permitted by
Apple in writing….The grants set forth in this License do not permit
you to, and you agree not to, install, use or run the Apple Font on any
non-Apple-branded hardware or to enable others to do so.
Notwithstanding the foregoing, you may use the Apple Font in
depictions, digital and/or print, of the Apple Watch, but only for
purposes of designing or developing applications for the Apple
Watch….(i) only one user may use the Apple Font at a time, and (ii) you
may not make the Apple Font available over a network where it could be
run or used by multiple computers at the same time. You may not rent,
lease, lend, trade, transfer, sell, sublicense or otherwise
redistribute the Apple Font in any unauthorized way.`